feat: expose a library surface and JSON snapshot API for headless consumers
abtop is binary-only today. This makes it a lib+bin crate so local tools can
reuse abtop's data-collection layer in-process — without reimplementing session
discovery, shelling out, or depending on the terminal frontend.

What changed:
- Split the entry point: src/main.rs becomes a thin shim that calls
  abtop::run(); all logic moves to src/lib.rs (a pure relocation).
- New App::tick_no_summaries(): refresh monitored data WITHOUT spawning
  `claude --print`, so headless consumers never touch the user's Claude quota.
  tick() is now tick_no_summaries() + drain_and_retry_summaries() — unchanged
  behavior for the TUI.
- New src/snapshot.rs: App::to_snapshot(interval_ms) returns a JSON-friendly
  Snapshot DTO (SessionView / McpServerView / ...). Pure read — never ticks
  or spawns.
- #[derive(Serialize)] on the model/host DTO types it serializes.
- New `abtop --json` flag: print one snapshot and exit.

Notes:
- The TUI, all existing flags, and output are unchanged.
- No new dependencies — serde and serde_json are already in the tree.
- snapshot.rs ships unit tests; all 164 tests + 1 doctest pass.
- Enum wire formats (SessionStatus as "Thinking"/..., chat roles as
  "user"/"assistant") are documented as stable.

Reference consumer: https://github.com/XKHoshizora/abtop-web-ui — a local-first
web dashboard built entirely on this API.
